py-chatbot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
transformers==4.46.1
faiss-cpu==1.12.0
pydantic==2.9.2
Pillow==10.4.0
tokenizers==0.20.1
torch==2.4.1
numpy==1.26.4
pymongo==4.6.0
requests==2.31.0
